Cleaning
I cleaned the file up a bit more so that it would be easier to see where things are. From what I can tell, our only problem left after this cleaning is that we need to make the "relations" method to work and it all should come together. In theory, this method should be the last piece of the puzzle and after this is comes the debugging process. As you can see I've made the structure for the relations method and all that's left is to implement the code.

Changes:
- Made ArrayList<Person> family as a class variable. So now we can access it from anywhere.
- Made revised the readFile method so that it takes in nextLine(). With this change the list can now accommodate for names such as 
   Henry VIII.
- Made the names NOT case sensitive. Although this provides more user accessibility, it is easier for the people coding to look through 
   the code without getting overloaded.

* Note: I think that we should focus on making the method for now and leave the small edits for later. Unless if we have two people who want to work on debugging while the other two works on the method which I think would be efficient.